fix: literal pathspecs for scratch unstage and split tests
Pass reserved scratch paths to git reset as :(literal) pathspecs so
basenames with backslash or glob metacharacters unstage instead of
escaping. Move reserved-scratch contract coverage out of gateway_test.go
to stay under the single-file test growth threshold.

Generated-By: looper 0.11.1 (runner=fixer, agent=grok-build)
